+++ libraries/node-core-library/src/FileSystem.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import * as pathUtilities from 'path';
+import * as nodeJsPath from 'path';
 import * as fs from 'fs';
 import * as fsx from 'fs-extra';
 
@@ -160,6 +160,34 @@ export interface IFileSystemCreateLinkOptions {
   newLinkPath: string;
 }
 
+const MOVE_DEFAULT_OPTIONS: Partial<IFileSystemMoveOptions> = {
+  overwrite: true,
+  ensureFolderExists: false
+};
+
+const READ_FOLDER_DEFAULT_OPTIONS: Partial<IFileSystemReadFolderOptions> = {
+  absolutePaths: false
+};
+
+const WRITE_FILE_DEFAULT_OPTIONS: Partial<IFileSystemWriteFileOptions> = {
+  ensureFolderExists: false,
+  convertLineEndings: undefined,
+  encoding: Encoding.Utf8
+};
+
+const APPEND_TO_FILE_DEFAULT_OPTIONS: Partial<IFileSystemWriteFileOptions> = {
+  ...WRITE_FILE_DEFAULT_OPTIONS
+};
+
+const READ_FILE_DEFAULT_OPTIONS: Partial<IFileSystemReadFileOptions> = {
+  encoding: Encoding.Utf8,
+  convertLineEndings: undefined
+};
+
+const DELETE_FILE_DEFAULT_OPTIONS: Partial<IFileSystemDeleteFileOptions> = {
+  throwIfNotExists: false
+};
+
 /**
  * The FileSystem API provides a complete set of recommended operations for interacting with the file system.
  *
@@ -208,6 +236,13 @@ export class FileSystem {
   }
 
   /**
+   * An async version of {@link FileSystem.getStatistics}.
+   */
+  public static getStatisticsAsync(path: string): Promise<fs.Stats> {
+    return fsx.stat(path);
+  }
+
+  /**
    * Updates the accessed and modified timestamps of the filesystem object referenced by path.
    * Behind the scenes it uses `fs.utimesSync()`.
    * The caller should specify both times in the `times` parameter.
@@ -215,9 +250,18 @@ export class FileSystem {
    * @param times - The times that the object should be updated to reflect.
    */
   public static updateTimes(path: string, times: IFileSystemUpdateTimeParameters): void {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    fsx.utimesSync(path, times.accessedTime as any, times.modifiedTime as any);
+    fsx.utimesSync(path, times.accessedTime, times.modifiedTime);
   }
+
+  /**
+   * An async version of {@link FileSystem.updateTimes}.
+   */
+  public static updateTimesAsync(path: string, times: IFileSystemUpdateTimeParameters): Promise<void> {
+    // This cast is needed because the fs-extra typings require both parameters
+    // to have the same type (number or Date), whereas Node.js does not require that.
+    return fsx.utimes(path, times.accessedTime as number, times.modifiedTime as number);
+  }
+
 
   /**
    * Changes the permissions (i.e. file mode bits) for a filesystem object.
@@ -230,12 +274,31 @@ export class FileSystem {
   }
 
   /**
+   * An async version of {@link FileSystem.changePosixModeBits}.
+   */
+  public static changePosixModeBitsAsync(path: string, mode: PosixModeBits): Promise<void> {
+    return fsx.chmod(path, mode);
+  }
+
+  /**
    * Retrieves the permissions (i.e. file mode bits) for a filesystem object.
    * Behind the scenes it uses `fs.chmodSync()`.
    * @param path - The absolute or relative path to the object that should be updated.
+   *
+   * @remarks
+   * This calls {@link FileSystem.getStatistics} to get the POSIX mode bits.
+   * If statistics in addition to the mode bits are needed, it is more efficient
+   * to call {@link FileSystem.getStatistics} directly instead.
    */
   public static getPosixModeBits(path: string): PosixModeBits {
     return FileSystem.getStatistics(path).mode;
+  }
+
+  /**
+   * An async version of {@link FileSystem.getPosixModeBits}.
+   */
+  public static async getPosixModeBitsAsync(path: string): Promise<PosixModeBits> {
+    return (await FileSystem.getStatisticsAsync(path)).mode;
   }
 
   /**
@@ -269,16 +332,51 @@ export class FileSystem {
    */
   public static move(options: IFileSystemMoveOptions): void {
     options = {
-      overwrite: true,
-      ensureFolderExists: false,
+      ...MOVE_DEFAULT_OPTIONS,
       ...options
     };
 
-    if (options.ensureFolderExists) {
-      FileSystem.ensureFolder(pathUtilities.basename(options.sourcePath));
-    }
+    try {
+      fsx.moveSync(options.sourcePath, options.destinationPath, { overwrite: options.overwrite });
+    } catch (error) {
+      if (options.ensureFolderExists) {
+        if (!FileSystem.isNotExistError(error)) {
+          throw error;
+        }
 
-    fsx.moveSync(options.sourcePath, options.destinationPath, { overwrite: options.overwrite });
+        const folderPath: string = nodeJsPath.dirname(options.destinationPath);
+        FileSystem.ensureFolder(folderPath);
+        fsx.moveSync(options.sourcePath, options.destinationPath, { overwrite: options.overwrite });
+      } else {
+        throw error;
+      }
+    }
+  }
+
+  /**
+   * An async version of {@link FileSystem.move}.
+   */
+  public static async moveAsync(options: IFileSystemMoveOptions): Promise<void> {
+    options = {
+      ...MOVE_DEFAULT_OPTIONS,
+      ...options
+    };
+
+    try {
+      await fsx.move(options.sourcePath, options.destinationPath, { overwrite: options.overwrite });
+    } catch (error) {
+      if (options.ensureFolderExists) {
+        if (!FileSystem.isNotExistError(error)) {
+          throw error;
+        }
+
+        const folderPath: string = nodeJsPath.dirname(options.destinationPath);
+        await FileSystem.ensureFolderAsync(nodeJsPath.dirname(folderPath));
+        await fsx.move(options.sourcePath, options.destinationPath, { overwrite: options.overwrite });
+      } else {
+        throw error;
+      }
+    }
   }
 
   // ===============
@@ -297,6 +395,13 @@ export class FileSystem {
   }
 
   /**
+   * An async version of {@link FileSystem.ensureFolder}.
+   */
+  public static ensureFolderAsync(folderPath: string): Promise<void> {
+    return fsx.ensureDir(folderPath);
+  }
+
+  /**
    * Reads the contents of the folder, not including "." or "..".
    * Behind the scenes it uses `fs.readdirSync()`.
    * @param folderPath - The absolute or relative path to the folder which should be read.
@@ -304,21 +409,55 @@ export class FileSystem {
    */
   public static readFolder(folderPath: string, options?: IFileSystemReadFolderOptions): string[] {
     options = {
-      absolutePaths: false,
+      ...READ_FOLDER_DEFAULT_OPTIONS,
       ...options
     };
 
-    if (!FileSystem.exists(folderPath)) {
-      throw new Error(`Folder does not exist: "${folderPath}"`);
+    let fileNames: string[]
+    try {
+      // @todo: Update this to use Node 10's `withFileTypes: true` option when we drop support for Node 8
+      fileNames = fsx.readdirSync(folderPath);
+    } catch (e) {
+      if (FileSystem.isNotExistError(e)) {
+        throw new Error(`Folder does not exist: "${folderPath}"`);
+      } else {
+        throw e;
+      }
     }
-
-    const fileNames: string[] = fsx.readdirSync(folderPath);
 
     if (options.absolutePaths) {
-      return fileNames.map(fileName => pathUtilities.resolve(folderPath, fileName));
+      return fileNames.map(fileName => nodeJsPath.resolve(folderPath, fileName));
+    } else {
+      return fileNames;
+    }
+  }
+
+  /**
+   * An async version of {@link FileSystem.readFolder}.
+   */
+  public static async readFolderAsync(folderPath: string, options?: IFileSystemReadFolderOptions): Promise<string[]> {
+    options = {
+      ...READ_FOLDER_DEFAULT_OPTIONS,
+      ...options
+    };
+
+    let fileNames: string[];
+    try {
+      // @todo: Update this to use Node 10's `withFileTypes: true` option when we drop support for Node 8
+      fileNames = await fsx.readdir(folderPath);
+    } catch (e) {
+      if (FileSystem.isNotExistError(e)) {
+        throw new Error(`Folder does not exist: "${folderPath}"`);
+      } else {
+        throw e;
+      }
     }
 
-    return fileNames;
+    if (options.absolutePaths) {
+      return fileNames.map(fileName => nodeJsPath.resolve(folderPath, fileName));
+    } else {
+      return fileNames;
+    }
   }
 
   /**
@@ -333,6 +472,13 @@ export class FileSystem {
   }
 
   /**
+   * An async version of {@link FileSystem.deleteFolder}.
+   */
+  public static deleteFolderAsync(folderPath: string): Promise<void> {
+    return fsx.remove(folderPath);
+  }
+
+  /**
    * Deletes the content of a folder, but not the folder itself. Also ensures the folder exists.
    * Behind the scenes it uses `fs-extra.emptyDirSync()`.
    * @remarks
@@ -342,6 +488,13 @@ export class FileSystem {
    */
   public static ensureEmptyFolder(folderPath: string): void {
     fsx.emptyDirSync(folderPath);
+  }
+
+  /**
+   * An async version of {@link FileSystem.ensureEmptyFolder}.
+   */
+  public static ensureEmptyFolderAsync(folderPath: string): Promise<void> {
+    return fsx.emptyDir(folderPath);
   }
 
   // ===============
@@ -359,22 +512,59 @@ export class FileSystem {
    */
   public static writeFile(filePath: string, contents: string | Buffer, options?: IFileSystemWriteFileOptions): void {
     options = {
-      ensureFolderExists: false,
-      convertLineEndings: undefined,
-      encoding: Encoding.Utf8,
+      ...WRITE_FILE_DEFAULT_OPTIONS,
       ...options
     };
-
-    if (options.ensureFolderExists) {
-      const folderPath: string = pathUtilities.dirname(filePath);
-      FileSystem.ensureFolder(folderPath);
-    }
 
     if (options.convertLineEndings) {
       contents = Text.convertTo(contents.toString(), options.convertLineEndings);
     }
 
-    fsx.writeFileSync(filePath, contents, { encoding: options.encoding });
+    try {
+      fsx.writeFileSync(filePath, contents, { encoding: options.encoding });
+    } catch (error) {
+      if (options.ensureFolderExists) {
+        if (!FileSystem.isNotExistError(error)) {
+          throw error;
+        }
+
+        const folderPath: string = nodeJsPath.dirname(filePath);
+        FileSystem.ensureFolder(folderPath);
+        fsx.writeFileSync(filePath, contents, { encoding: options.encoding });
+      } else {
+        throw error;
+      }
+    }
+  }
+
+  /**
+   * An async version of {@link FileSystem.writeFile}.
+   */
+  public static async writeFileAsync(filePath: string, contents: string | Buffer, options?: IFileSystemWriteFileOptions): Promise<void> {
+    options = {
+      ...WRITE_FILE_DEFAULT_OPTIONS,
+      ...options
+    };
+
+    if (options.convertLineEndings) {
+      contents = Text.convertTo(contents.toString(), options.convertLineEndings);
+    }
+
+    try {
+      await fsx.writeFile(filePath, contents, { encoding: options.encoding });
+    } catch (error) {
+      if (options.ensureFolderExists) {
+        if (!FileSystem.isNotExistError(error)) {
+          throw error;
+        }
+
+        const folderPath: string = nodeJsPath.dirname(filePath);
+        await FileSystem.ensureFolderAsync(folderPath);
+        await fsx.writeFile(filePath, contents, { encoding: options.encoding });
+      } else {
+        throw error;
+      }
+    }
   }
 
   /**
@@ -388,22 +578,59 @@ export class FileSystem {
    */
   public static appendToFile(filePath: string, contents: string | Buffer, options?: IFileSystemWriteFileOptions): void {
     options = {
-      ensureFolderExists: false,
-      convertLineEndings: undefined,
-      encoding: Encoding.Utf8,
+      ...APPEND_TO_FILE_DEFAULT_OPTIONS,
       ...options
     };
-
-    if (options.ensureFolderExists) {
-      const folderPath: string = pathUtilities.dirname(filePath);
-      FileSystem.ensureFolder(folderPath);
-    }
 
     if (options.convertLineEndings) {
       contents = Text.convertTo(contents.toString(), options.convertLineEndings);
     }
 
-    fsx.appendFileSync(filePath, contents, { encoding: options.encoding });
+    try {
+      fsx.appendFileSync(filePath, contents, { encoding: options.encoding });
+    } catch (error) {
+      if (options.ensureFolderExists) {
+        if (!FileSystem.isNotExistError(error)) {
+          throw error;
+        }
+
+        const folderPath: string = nodeJsPath.dirname(filePath);
+        FileSystem.ensureFolder(folderPath);
+        fsx.appendFileSync(filePath, contents, { encoding: options.encoding });
+      } else {
+        throw error;
+      }
+    }
+  }
+
+  /**
+   * An async version of {@link FileSystem.appendToFile}.
+   */
+  public static async appendToFileAsync(filePath: string, contents: string | Buffer, options?: IFileSystemWriteFileOptions): Promise<void> {
+    options = {
+      ...APPEND_TO_FILE_DEFAULT_OPTIONS,
+      ...options
+    };
+
+    if (options.convertLineEndings) {
+      contents = Text.convertTo(contents.toString(), options.convertLineEndings);
+    }
+
+    try {
+      await fsx.appendFile(filePath, contents, { encoding: options.encoding });
+    } catch (error) {
+      if (options.ensureFolderExists) {
+        if (!FileSystem.isNotExistError(error)) {
+          throw error;
+        }
+
+        const folderPath: string = nodeJsPath.dirname(filePath);
+        await FileSystem.ensureFolderAsync(folderPath);
+        await fsx.appendFile(filePath, contents, { encoding: options.encoding });
+      } else {
+        throw error;
+      }
+    }
   }
 
   /**
@@ -414,8 +641,7 @@ export class FileSystem {
    */
   public static readFile(filePath: string, options?: IFileSystemReadFileOptions): string {
     options = {
-      encoding: Encoding.Utf8,
-      convertLineEndings: undefined,
+      ...READ_FILE_DEFAULT_OPTIONS,
       ...options
     };
 
@@ -423,6 +649,24 @@ export class FileSystem {
     if (options.convertLineEndings) {
       contents = Text.convertTo(contents, options.convertLineEndings);
     }
+
+    return contents;
+  }
+
+  /**
+   * An async version of {@link FileSystem.readFile}.
+   */
+  public static async readFileAsync(filePath: string, options?: IFileSystemReadFileOptions): Promise<string> {
+    options = {
+      ...READ_FILE_DEFAULT_OPTIONS,
+      ...options
+    };
+
+    let contents: string = (await FileSystem.readFileToBufferAsync(filePath)).toString(options.encoding);
+    if (options.convertLineEndings) {
+      contents = Text.convertTo(contents, options.convertLineEndings);
+    }
+
     return contents;
   }
 
@@ -436,12 +680,26 @@ export class FileSystem {
   }
 
   /**
+   * An async version of {@link FileSystem.readFileToBuffer}.
+   */
+  public static readFileToBufferAsync(filePath: string): Promise<Buffer> {
+    return fsx.readFile(filePath);
+  }
+
+  /**
    * Copies a file from one location to another.
    * By default, destinationPath is overwritten if it already exists.
    * Behind the scenes it uses `fs.copyFileSync()`.
    */
   public static copyFile(options: IFileSystemCopyFileOptions): void {
     fsx.copySync(options.sourcePath, options.destinationPath);
+  }
+
+  /**
+   * An async version of {@link FileSystem.copyFile}.
+   */
+  public static async copyFileAsync(options: IFileSystemCopyFileOptions): Promise<void> {
+    await fsx.copy(options.sourcePath, options.destinationPath);
   }
 
   /**
@@ -452,17 +710,33 @@ export class FileSystem {
    */
   public static deleteFile(filePath: string, options?: IFileSystemDeleteFileOptions): void {
     options = {
-      throwIfNotExists: false,
+      ...DELETE_FILE_DEFAULT_OPTIONS,
       ...options
     };
 
-    if (options.throwIfNotExists) {
+    try {
       fsx.unlinkSync(filePath);
-    } else {
-      try {
-        fsx.unlinkSync(filePath);
-      } catch (error) {
-        /* no-op */
+    } catch (error) {
+      if (options.throwIfNotExists || !FileSystem.isNotExistError(error)) {
+        throw error;
+      }
+    }
+  }
+
+  /**
+   * An async version of {@link FileSystem.deleteFile}.
+   */
+  public static async deleteFileAsync(filePath: string, options?: IFileSystemDeleteFileOptions): Promise<void> {
+    options = {
+      ...DELETE_FILE_DEFAULT_OPTIONS,
+      ...options
+    };
+
+    try {
+      await fsx.unlink(filePath);
+    } catch (error) {
+      if (options.throwIfNotExists || !FileSystem.isNotExistError(error)) {
+        throw error;
       }
     }
   }
@@ -481,12 +755,27 @@ export class FileSystem {
   }
 
   /**
+   * An async version of {@link FileSystem.getLinkStatistics}.
+   */
+  public static getLinkStatisticsAsync(path: string): Promise<fs.Stats> {
+    return fsx.lstat(path);
+  }
+
+  /**
    * Creates a Windows "directory junction". Behaves like `createSymbolicLinkToFile()` on other platforms.
    * Behind the scenes it uses `fs.symlinkSync()`.
    */
   public static createSymbolicLinkJunction(options: IFileSystemCreateLinkOptions): void {
     // For directories, we use a Windows "junction".  On POSIX operating systems, this produces a regular symlink.
     fsx.symlinkSync(options.linkTargetPath, options.newLinkPath, 'junction');
+  }
+
+  /**
+   * An async version of {@link FileSystem.createSymbolicLinkJunction}.
+   */
+  public static createSymbolicLinkJunctionAsync(options: IFileSystemCreateLinkOptions): Promise<void> {
+    // For directories, we use a Windows "junction".  On POSIX operating systems, this produces a regular symlink.
+    return fsx.symlink(options.linkTargetPath, options.newLinkPath, 'junction');
   }
 
   /**
@@ -498,11 +787,25 @@ export class FileSystem {
   }
 
   /**
+   * An async version of {@link FileSystem.createSymbolicLinkFile}.
+   */
+  public static createSymbolicLinkFileAsync(options: IFileSystemCreateLinkOptions): Promise<void> {
+    return fsx.symlink(options.linkTargetPath, options.newLinkPath, 'file');
+  }
+
+  /**
    * Creates a symbolic link to a folder (on Windows this requires elevated permissionsBits).
    * Behind the scenes it uses `fs.symlinkSync()`.
    */
   public static createSymbolicLinkFolder(options: IFileSystemCreateLinkOptions): void {
     fsx.symlinkSync(options.linkTargetPath, options.newLinkPath, 'dir');
+  }
+
+  /**
+   * An async version of {@link FileSystem.createSymbolicLinkFolder}.
+   */
+  public static createSymbolicLinkFolderAsync(options: IFileSystemCreateLinkOptions): Promise<void> {
+    return fsx.symlink(options.linkTargetPath, options.newLinkPath, 'dir');
   }
 
   /**
@@ -514,11 +817,37 @@ export class FileSystem {
   }
 
   /**
+   * An async version of {@link FileSystem.createHardLink}.
+   */
+  public static createHardLinkAsync(options: IFileSystemCreateLinkOptions): Promise<void> {
+    return fsx.link(options.linkTargetPath, options.newLinkPath);
+  }
+
+  /**
    * Follows a link to its destination and returns the absolute path to the final target of the link.
    * Behind the scenes it uses `fs.realpathSync()`.
    * @param linkPath - The path to the link.
    */
   public static getRealPath(linkPath: string): string {
     return fsx.realpathSync(linkPath);
+  }
+
+  /**
+   * An async version of {@link FileSystem.getRealPath}.
+   */
+  public static getRealPathAsync(linkPath: string): Promise<string> {
+    return fsx.realpath(linkPath);
+  }
+
+  // ===============
+  // UTILITY FUNCTIONS
+  // ===============
+
+  /**
+   * Returns true if the error provided indicates the file or folder
+   * does not exist.
+   */
+  public static isNotExistError(error: NodeJS.ErrnoException): boolean {
+    return error.code === 'ENOENT' || error.code === 'ENOTDIR';
   }
 }