+++ core-build/gulp-core-build-mocha/CHANGELOG.md
@@ -1,6 +1,21 @@
 # Change Log - @microsoft/gulp-core-build-mocha
 
-This log was last generated on Sat, 02 May 2020 00:08:16 GMT and should not be manually modified.
+This log was last generated on Fri, 22 May 2020 15:08:43 GMT and should not be manually modified.
+
+## 3.8.12
+Fri, 22 May 2020 15:08:43 GMT
+
+*Version update only*
+
+## 3.8.11
+Thu, 21 May 2020 23:09:44 GMT
+
+*Version update only*
+
+## 3.8.10
+Thu, 21 May 2020 15:42:00 GMT
+
+*Version update only*
 
 ## 3.8.9
 Sat, 02 May 2020 00:08:16 GMT