+++ core-build/gulp-core-build-serve/CHANGELOG.md
@@ -1,6 +1,31 @@
 # Change Log - @microsoft/gulp-core-build-serve
 
-This log was last generated on Wed, 06 May 2020 08:23:45 GMT and should not be manually modified.
+This log was last generated on Fri, 22 May 2020 15:08:42 GMT and should not be manually modified.
+
+## 3.7.14
+Fri, 22 May 2020 15:08:42 GMT
+
+*Version update only*
+
+## 3.7.13
+Thu, 21 May 2020 23:09:44 GMT
+
+*Version update only*
+
+## 3.7.12
+Thu, 21 May 2020 15:41:59 GMT
+
+*Version update only*
+
+## 3.7.11
+Tue, 19 May 2020 15:08:19 GMT
+
+*Version update only*
+
+## 3.7.10
+Fri, 15 May 2020 08:10:59 GMT
+
+*Version update only*
 
 ## 3.7.9
 Wed, 06 May 2020 08:23:45 GMT