+++ webpack/loader-load-themed-styles/CHANGELOG.md
@@ -1,6 +1,31 @@
 # Change Log - @microsoft/loader-load-themed-styles
 
-This log was last generated on Wed, 06 May 2020 08:23:45 GMT and should not be manually modified.
+This log was last generated on Fri, 22 May 2020 15:08:42 GMT and should not be manually modified.
+
+## 1.8.17
+Fri, 22 May 2020 15:08:42 GMT
+
+*Version update only*
+
+## 1.8.16
+Thu, 21 May 2020 23:09:44 GMT
+
+*Version update only*
+
+## 1.8.15
+Thu, 21 May 2020 15:42:00 GMT
+
+*Version update only*
+
+## 1.8.14
+Tue, 19 May 2020 15:08:19 GMT
+
+*Version update only*
+
+## 1.8.13
+Fri, 15 May 2020 08:10:59 GMT
+
+*Version update only*
 
 ## 1.8.12
 Wed, 06 May 2020 08:23:45 GMT